--- a/fdbserver/ProxyCommitData.actor.h
+++ b/fdbserver/ProxyCommitData.actor.h
@@ -227,6 +227,7 @@ struct ProxyCommitData {
 	KeyRangeMap<std::set<Key>> vecBackupKeys;
 	uint64_t commitVersionRequestNumber;
 	uint64_t mostRecentProcessedRequestNumber;
+	PromiseStream<Void> committedBatches; // for notification commitBatcher
 	KeyRangeMap<Deque<std::pair<Version, int>>> keyResolvers;
 	// When all resolvers process system keys (for private mutations), the "keyResolvers"
 	// only tracks normalKeys. This is used for tracking versions for systemKeys.